Force custom catalog for NFD if specified
Currently we only fall back to a custom catalog source when the
NFD operator is not present in the default catalog. These changes
will force using the custom catalog - if specified - instead of the
default one. This is needed in cases when the default operator is
broken, so we can carry on testing the GPU/NNO operator until the
default NFD operator is fixed.